% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexcal.R
\name{fi_index}
\alias{fi_index}
\title{Title}
\usage{
fi_index(
  x,
  species,
  code,
  runs,
  slope = "Actual.river.slope",
  tempjan = "Temp.jan",
  tempjul = "Temp.jul",
  water = "Water.source.type",
  sediment = "Natural.sediment",
  geomorph = "Geomorph.river.type",
  method = "Method",
  area = "Area.ctch",
  floodpsite = "Floodplain.site",
  distsource = "Distance.from.source",
  labels,
  type,
  labels2,
  type2,
  len_up = "Number.length.over.150",
  len_down = "Number.length.below.150",
  tolerance = 1e-07
)
}
\arguments{
\item{tolerance}{}
}
\description{
Title
}
