% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fi_observed}
\alias{fi_observed}
\title{Title}
\usage{
fi_observed(
  xdf,
  spp,
  labels = NULL,
  type = NULL,
  labels2 = NULL,
  type2 = NULL,
  len = FALSE,
  tolerance
)
}
\arguments{
\item{tolerance}{}
}
\description{
Title
}
