% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metricscore}
\alias{metricscore}
\title{Standardized distance between the observed (sampled) and expected (no alteration) values in the community}
\usage{
metricscore(
  env,
  lat,
  labels,
  region,
  medtype,
  mode = "none",
  observed,
  expected,
  obslength,
  efitype
)
}
\arguments{
\item{efitype}{}
}
\description{
Standardized distance between the observed (sampled) and expected (no alteration) values in the community
}
