% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexcal.R
\name{fi_index}
\alias{fi_index}
\title{Title}
\usage{
fi_index(
  x,
  species,
  code,
  runs,
  slope,
  tempjan,
  tempjul,
  water,
  sediment,
  geomorph,
  method,
  area,
  floodpsite,
  distsource,
  labels,
  type,
  labels2,
  type2,
  len_up,
  len_down,
  tolerance = 1e-07
)
}
\arguments{
\item{tolerance}{}
}
\description{
Title
}
